# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Flyff Idle API
  version: 0.1.0
  description: API schema for auth, characters, inventory, admin actions, and game data.
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/characters:
    get:
      tags:
        - Characters
      summary: List characters for the authenticated player
      operationId: listCharacters
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Character roster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - Characters
      summary: Create a character in an open roster slot
      operationId: createCharacter
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCharacterRequest"
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Character slot already occupied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/characters/{characterId}:
    get:
      tags:
        - Characters
      summary: Get one character owned by the authenticated player
      operationId: getCharacter
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
      responses:
        "200":
          description: Character details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Characters
      summary: Delete a character after name confirmation
      operationId: deleteCharacter
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
        - in: query
          name: confirmationName
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "204":
          description: Character deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/characters/{characterId}/inventory/{slotIndex}/equip:
    post:
      tags:
        - Characters
      summary: Equip an inventory item into an equipment set
      operationId: equipCharacterItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
        - $ref: "#/components/parameters/SlotIndexPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentSetRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/characters/{characterId}/equipment/{equipmentSlot}/unequip:
    post:
      tags:
        - Characters
      summary: Unequip an item from an equipment set
      operationId: unequipCharacterItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
        - $ref: "#/components/parameters/EquipmentSlotPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentSetRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/characters/{characterId}/equipment/ammo/consume:
    post:
      tags:
        - Characters
      summary: Consume one arrow from an equipment set
      operationId: consumeCharacterAmmo
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentSetRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/characters/{characterId}/consumables/{resource}:
    post:
      tags:
        - Characters
      summary: Equip or clear a character consumable slot
      operationId: setCharacterConsumable
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
        - $ref: "#/components/parameters/ConsumableResourcePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipConsumableRequest"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/characters/{characterId}/consumables/{resource}/consume:
    post:
      tags:
        - Characters
      summary: Consume one item from a character consumable slot
      operationId: consumeCharacterConsumable
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CharacterIdPath"
        - $ref: "#/components/parameters/ConsumableResourcePath"
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    CharacterIdPath:
      in: path
      name: characterId
      required: true
      schema:
        type: string
    SlotIndexPath:
      in: path
      name: slotIndex
      required: true
      schema:
        type: integer
        minimum: 0
        maximum: 99
    EquipmentSlotPath:
      in: path
      name: equipmentSlot
      required: true
      schema:
        $ref: "#/components/schemas/CharacterEquipmentSlot"
...
    ConsumableResourcePath:
      in: path
      name: resource
      required: true
      schema:
        $ref: "#/components/schemas/CharacterConsumableResource"
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid bearer token, or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: The operation failed unexpectedly
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CharacterListResponse:
      type: object
      required:
        - characters
      properties:
        characters:
          type: array
          items:
            $ref: "#/components/schemas/Character"
    CharacterResponse:
      type: object
      required:
        - character
      properties:
        character:
          $ref: "#/components/schemas/Character"
    CreateCharacterRequest:
      type: object
      required:
        - slotIndex
        - name
        - gender
      properties:
        slotIndex:
          type: integer
          minimum: 0
          maximum: 7
          example: 2
        name:
          type: string
          minLength: 3
          maxLength: 16
          pattern: "^[A-Za-z][A-Za-z0-9 ]*$"
          example: FlarisHero
        gender:
          $ref: "#/components/schemas/CharacterGender"
...
    EquipmentSetRequest:
      type: object
      properties:
        equipmentSet:
          type: integer
          minimum: 0
          maximum: 2
          default: 0
...
    EquipConsumableRequest:
      type: object
      required:
        - slotIndex
      properties:
        slotIndex:
          type: integer
          minimum: 0
          maximum: 99
          nullable: true
...
    Character:
      type: object
      required:
        - id
        - slotIndex
        - name
        - gender
        - job
        - progressionRank
        - location
        - level
        - exp
        - penya
        - stats
        - skillLevels
        - consumableLoadout
        - equipment
        - equipmentSets
        - ammoQuantity
        - ammoQuantities
        - activeQuestIds
        - completedQuestIds
        - inventory
      properties:
        id:
          type: string
        slotIndex:
          type: integer
          minimum: 0
          maximum: 7
        name:
          type: string
        gender:
          $ref: "#/components/schemas/CharacterGender"
        job:
          type: string
        progressionRank:
          $ref: "#/components/schemas/CharacterProgressionRank"
        location:
          type: string
        level:
          type: integer
          minimum: 1
        exp:
          type: integer
          minimum: 0
        penya:
          type: integer
          minimum: 0
        stats:
          $ref: "#/components/schemas/CharacterStats"
        skillLevels:
          $ref: "#/components/schemas/CharacterSkillLevels"
        consumableLoadout:
          $ref: "#/components/schemas/CharacterConsumableLoadout"
        equipment:
          $ref: "#/components/schemas/CharacterEquipment"
        equipmentSets:
          type: array
          minItems: 3
          maxItems: 3
          items:
            $ref: "#/components/schemas/CharacterEquipment"
        ammoQuantity:
          type: integer
          minimum: 0
        ammoQuantities:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: integer
            minimum: 0
        activeQuestIds:
          type: array
          items:
            type: integer
            minimum: 1
        completedQuestIds:
          type: array
          items:
            type: integer
            minimum: 1
        inventory:
          $ref: "#/components/schemas/CharacterInventory"
    CharacterGender:
      type: string
      enum:
        - male
        - female
    CharacterProgressionRank:
      type: string
      enum:
        - normal
        - master
        - hero
    CharacterStats:
      type: object
      required:
        - str
        - sta
        - dex
        - int
      properties:
        str:
          type: integer
          minimum: 15
          maximum: 999
          example: 15
        sta:
          type: integer
          minimum: 15
          maximum: 999
          example: 15
        dex:
          type: integer
          minimum: 15
          maximum: 999
          example: 15
        int:
          type: integer
          minimum: 15
          maximum: 999
          example: 15
    CharacterSkillLevels:
      type: object
      additionalProperties:
        type: integer
        minimum: 0
        maximum: 20
      example:
        vagrant-clean-hit: 1
    CharacterConsumableResource:
      type: string
      enum:
        - hp
        - mp
        - fp
    CharacterConsumableSlot:
      type: object
      nullable: true
      required:
        - itemId
        - quantity
      properties:
        itemId:
          type: string
        quantity:
          type: integer
          minimum: 1
    CharacterConsumableLoadout:
      type: object
      required:
        - hp
        - mp
        - fp
      properties:
        hp:
          $ref: "#/components/schemas/CharacterConsumableSlot"
        mp:
          $ref: "#/components/schemas/CharacterConsumableSlot"
        fp:
          $ref: "#/components/schemas/CharacterConsumableSlot"
    CharacterEquipment:
      type: object
      required:
        - helmet
        - suit
        - gloves
        - boots
        - flying
        - csBoots
        - csGloves
        - csSuit
        - csHelm
        - mask
        - cloak
        - ammo
        - offhand
        - mainhand
        - ringR
        - earringR
        - necklace
        - earringL
        - ringL
      properties:
        helmet:
          type: string
          nullable: true
        suit:
          type: string
          nullable: true
        gloves:
          type: string
          nullable: true
        boots:
          type: string
          nullable: true
        flying:
          type: string
          nullable: true
        csBoots:
          type: string
          nullable: true
        csGloves:
          type: string
          nullable: true
        csSuit:
          type: string
          nullable: true
        csHelm:
          type: string
          nullable: true
        mask:
          type: string
          nullable: true
        cloak:
          type: string
          nullable: true
        ammo:
          type: string
          nullable: true
        offhand:
          type: string
          nullable: true
        mainhand:
          type: string
          nullable: true
        ringR:
          type: string
          nullable: true
        earringR:
          type: string
          nullable: true
        necklace:
          type: string
          nullable: true
        earringL:
          type: string
          nullable: true
        ringL:
          type: string
          nullable: true
    CharacterEquipmentSlot:
      type: string
      enum:
        - helmet
        - suit
        - gloves
        - boots
        - flying
        - csBoots
        - csGloves
        - csSuit
        - csHelm
        - mask
        - cloak
        - ammo
        - offhand
        - mainhand
        - ringR
        - earringR
        - necklace
        - earringL
        - ringL
    CharacterInventory:
      type: object
      required:
        - size
        - items
      properties:
        size:
          type: integer
          minimum: 1
          example: 50
        items:
          type: array
          items:
            $ref: "#/components/schemas/CharacterInventoryItem"
    CharacterInventoryItem:
      type: object
      required:
        - slotIndex
        - itemId
        - quantity
      properties:
        slotIndex:
          type: integer
          minimum: 0
        itemId:
          type: string
        quantity:
          type: integer
          minimum: 1
...
    ErrorResponse:
      type: object
      required:
        - code
        - error
      properties:
        code:
          type: string
          enum:
            - authentication_required
            - conflict
            - domain_rule_failed
            - forbidden
            - internal_error
            - invalid_request
            - not_found
            - upstream_error
        error:
          type: string